Select genes by name instead of numerical index
It turns out that in version 0.20, pandas changed the way that index
merging happens [1]_, and now preserves the order of the *left* index,
rather than the right one. We could change the order of index merging,
but then people with pandas 0.19 would have issues. Better to make
ourselves independent of index sort order, and grab the exact genes
we want independently.

.. [1]: https://pandas.pydata.org/pandas-docs/version/0.20/whatsnew.html#index-intersection-and-inner-join-now-preserve-the-order-of-the-left-index