Fixed bug in import alias inconsistency of ndimage in ch. 3
The first code cell in ch. 3 is explanatory only, but if the user
tried to run the code they would get an error because the import
statement aliases ndimage to nd but the subsequent calls to the
module use ndi.

This occurs in the final code cells as well, but the exception is
avoided there because there is one cell where ndimage is imported
as ndi.

This commit fixes the issue by making the ndimage import aliasing
consistent throughout the document."